Extract task enqueue pipeline out of ContextImpl
The #enqueueTasks and #encodeTaskMessage methods made ContextImpl
oversized, so move the handle validation, deduplication planning,
payload encoding, and queue dispatch into a new tasks/enqueue.ts
module.  ContextImpl now delegates to enqueueTasks(), passing only
the small slice of itself (federation, codec, origin, data) the
pipeline needs.

Pull the shared task-test helpers (the schema factory, stock schemas,
base federation options, and the recording MockQueue) into a new
testing/mq-tasks.ts module, and split the enqueue-specific cases out
of tasks.test.ts into enqueue.test.ts.

Teach the fixture-usage check to expand glob patterns in its
allowlist so the whole testing/ directory is covered by a single
entry instead of one path per file.
