+++ packages/fedify/src/testing/mod.ts
@@ -6,11 +6,14 @@ export {
 export {
   type Envelope,
   envelopeSchema,
-  makeSchema,
-  MockQueue,
   type MockQueueOptions,
-  numberSchema,
-  stringSchema,
 } from "./tasks.ts";
 // Without the export below, `test:cfworkers` makes an error.
 export { testDefinitions } from "@fedify/fixture";
+export {
+  baseOptions,
+  makeSchema,
+  MockQueue,
+  numberSchema,
+  stringSchema,
+} from "./mq-tasks.ts";
